fill out core implementation, rename some bits
The word 'tile' has a special meaning for factorio as a thing that
is placed on the floor, so we changed the name 'TileSet' to 'Palette'
to avoid confusion there.

We started filling out the Palette typeclass, guided by what proves
to be required as we start implementing imageToJson and friends.

We also started implementing the Palette typeclass for two options
available in the base game - purely flooring, or all monochrome
options (flooring + walls + gates).