rework executable for previewing & debugging
We now have an interface where we can ask for a preview of the
output image, or print out the blueprint to stdout as a string (ready
for importing straight into the game) or as a JSON (intended for
debugging usage). There is also the ability to specify which
tileset/palette to use - currently two options, 'flooring' or 'all'.